Drop Fragment, reimplement MediatorBuilder
